atc: behaviour: add /pipelines/:pipeline_id endpoints
All of the API tests are still only testing the old endpoints.

The recently added RouteAction function simply strips off the
"ByPipelineID" suffix from routes to determine the action - not pretty,
but it works for now.

The pagination endpoints also return the "ByPipelineID" endpoints in the
`Link` header - this will break go-concourse/fly as go-concourse ignores
the actual URL from the `Link` header, and only considers the query
parameters. This'll be addressed in a later commit.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>